[Tests] Move UI test to Core Unit Test
